Refactor component styles and imports
