+++ src/index.css
@@ -1,7 +1,3 @@
-@tailwind base;
-@tailwind components;
-@tailwind utilities;
-
 :root {
   font-family: Inter, system-ui, Avenir, Helvetica, Arial, sans-serif;
   line-height: 1.5;
@@ -96,32 +92,11 @@ body {
   color: #01a043;
 }
 
-.checkoutCard {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  width: 85%;
-  border: var(--dnb_seaGreen) solid 1px;
-  background-color: var(--dnb_mintGreen);
-  border-radius: 5px;
-  padding: 1rem;
-  margin-bottom: 1rem;
-  margin-top: 0.5rem;
-  color: black;
-  box-shadow: 0px 2px 4px rgba(0, 0, 0, 0.497);
-}
-
 p {
   color: black;
 }
 
-.dottedBottom {
-  border-bottom: dotted var(--dnb_mintGreen) 3px;
-  padding: 1rem;
-}
-
-.floorCardText {
+.textMargin {
   margin-left: 0.25rem;
   margin-right: 0.25rem;
 }
@@ -129,20 +104,4 @@ p {
 .reverse {
   display: flex;
   flex-direction: column-reverse;
-}
-
-.pickSpotCard {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  width: 85%;
-  border: var(--dnb_seaGreen) solid 1px;
-  background-color: var(--dnb_mintGreen);
-  border-radius: 5px;
-  padding: 1rem;
-  margin-bottom: 1rem;
-  margin-top: 0.5rem;
-  color: black;
-  box-shadow: 0px 2px 4px rgba(0, 0, 0, 0.497);
 }
