+++ src/pages/AdminPage.tsx
@@ -2,7 +2,7 @@ import parkData from "../data/parkingData.json";
 import RatesDisplay from "../components/RatesDisplay";
 import RatesForm from "../components/RatesForm";
 import CapacityDisplay from "../components/capacityDisplay/CapacityDisplay";
-import CapacityForm from "../components/CapacityForm";
+import CapacityForm from "../components/capacityForm/CapacityForm";
 import { useState } from "react";
 import { Anchor } from "@dnb/eufemia/components";
 import { Spot } from "../types/types";
